requirements.txt
pygame~=2.1.3
gym~=0.26.2
numpy~=1.25.2
pillow~=8.4.0
matplotlib~=3.4.3